Refactor classifier validators
